requirement.txt
asgiref==3.3.4
boto3==1.17.99
botocore==1.20.99
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-common-helpers==0.9.2
django-cors-headers==3.7.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
django-storages==1.11.1
idna==2.10
jmespath==0.10.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-http-client==3.3.2
pytz==2021.1
requests==2.25.1
s3transfer==0.4.2
sendgrid==6.7.1
sendgrid-python==0.1.1
six==1.16.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
urllib3==1.26.5
